Align with Faust implementation
Once faust itself is delivered with the toolkit, a lot of this project won't be needed!

That said, we'll keep it up to date to ensure that those with an older version of Faust
can still have the tools.

With THIS commit it is not necessary to install the kit on the pedal, though that is
still an option.